Back-end/requirements.txt
Flask
firebase-admin
pyrebase4
Werkzeug
